move masked circuits tests to separate file
